氷点下の CPU 温度が返ると符号なし整数への変換エラーになるのを修正
本当にデバイスファイルが返しうるのかは不明だが
安全側に倒しておく

+++ rshanghai/src/sysmod/health.rs
@@ -251,7 +251,7 @@ async fn get_cpu_temp() -> Result<CpuTemp> {
     let text = String::from_utf8_lossy(&buf);
 
     // 'C を 1000 倍した整数が得られるので変換する
-    let temp: u64 = text.trim().parse()?;
+    let temp: f64 = text.trim().parse()?;
     let temp = temp as f64 / 1000.0;
 
     Ok(CpuTemp { temp })
